requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-dotenv==1.0.1
guardrails-ai==0.5.10
httpx==0.27.2
tenacity==8.3.0
streamlit==1.39.0

